Create IsValidSolution.py
After receiving a string s that only contains parentheses, brackets, and curly brackets, determine if the input is valid. Return your answer in the form of a boolean data type. The input is only deemed valid if all of the parentheses, brackets, and curly brackets, open and close in order. For example: "( )", "[ ]", "{ }", "{ ( ) }", are all valid inputs. "[ { ] }" is NOT a valid input.