Create BaseballGameSolution.py
Write a program that parses a string for operation instructions. In your program, complete the operations specified by the user. Operations include: "C" to clear the last score passed, "D" to double the last score passed, "+" to add the two previous scores passed, or some integer x to store an operand for future use. Return the sum of all of the scores after all operations are applied.