Create RemoveElement.py
Write a program that takes in an array, nums, and a target value, val. Remove all instances of the target value in the nums array. Return how many elements in nums do NOT equal the target value. This procedure only modifies the input array. Creating a new array is not allowed.